+++ linalg/assign_mul_bra_add_mul_ket_add_bi.c
@@ -30,7 +30,7 @@ void assign_mul_bra_add_mul_ket_add_bi(bispinor * const R, bispinor * const S, b
   double a1,b1,a2,b2;
 #if ( defined SSE || defined SSE2 )
    w_=calloc(N+1, sizeof(spinor));
-   w = (spinor *)(((unsigned int)(w_)+ALIGN_BASE)&~ALIGN_BASE);
+   w = (spinor *)(((unsigned long int)(w_)+ALIGN_BASE)&~ALIGN_BASE);
 #else
    w_=calloc(N, sizeof(spinor));
    w = w_;
